@@ -1,4 +1,4 @@
-class AddAttachmentPictureToUsers < ActiveRecord::Migration
+class AddAttachmentPictureToUsers < ActiveRecord::Migration[5.0]
   def self.up
     change_table :users do |t|
       t.attachment :picture
